Add new batching submodule & edit web-app example project (#14)
* Add new batching submodule core

* Refactor Serilog TelegramSink to utilize BatchCycleManager with configurable rules

The Serilog TelegramSink was refactored to utilize a new BatchCycleManager. This manager is configured with a set of rules (such as BatchSizeRule and OncePerTimeRule) which specify when batches of log messages should be emitted. Also, the LogsQueueAccessContext was updated to use a ChannelReader for non-blocking, thread-safe queue operations. This change aims to improve the performance of the TelegramSink by allowing more fine-grained control over batch processing and improving thread safety.

* Optimized DateTime Now call in OncePerTimeRule

Moved direct call of DateTime.Now function to using declaration for improvement of code readability and performance. This change minimizes the call stack depth and reduces potential temporal coupling issue.

* Refactor TelegramSink for better batch processing

The class `TelegramSink` has been refactored for better handling of batch processing. The implementation now supports both `IDisposable` and `IAsyncDisposable` interfaces for graceful disposing of batch cycles. The flushing of remaining messages has been extracted into a separate method and is achieved asynchronously when disposing. Batch processing is now handled in the `EmitBatchInternalAsync` method which improves code readability and provides clearer separation of concerns. In addition, the `BatchCycleManager` now considers sink configuration rule-checking period rather than a hardcoded value. These changes were made to improve batching efficiency and ensure reliability when disposing.

* Refactor Telegram Logging Configuration and Fix Log Formatter

Deleted unused Telegram extensions and moved Telegram configuration to the main project to simplify code structure. This makes it easier to maintain and update configurations in the future.

Also fixed a negation issue in the `DefaultLogFormatter` class that was causing incorrect logic when determining if `logEntries` was empty or not. This resolves potential issues of incorrect log formatting.

* Remove LogsGenerator, integrate logging into Program.cs

Removed LogsGenerator controller as it was unnecessarily complicating the logging process. Integrated log handling into Program.cs for simplified, consolidated code structure. Updated Serilog.AspNetCore version to 7.0.0 improving the efficiency of the application. Changed project reference from X.Serilog.Sinks.Telegram.Extensions to X.Serilog.Sinks.Telegram to reflect current project structure.

---------

Co-authored-by: Vladyslav Bardin <vbardin-ecw@cricut.com>